+++ support/chip-testing/test/app-fast/BOOL.test.ts
@@ -5,5 +5,5 @@
  */
 
 describe("BOOL", () => {
-    chip("BOOL/*");
+    chip("BOOL/**");
 });