+++ support/chip-testing/test/app-fast/MWOM.test.ts
@@ -5,5 +5,5 @@
  */
 
 describe("MWOM", () => {
-    chip("MWOM/*");
+    chip("MWOM/**");
 });